+++ docs/html/search/classes_3.js
@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['linkentrybroker_303',['LinkEntryBroker',['../classace__time_1_1extended_1_1LinkEntryBroker.html',1,'ace_time::extended::LinkEntryBroker'],['../classace__time_1_1basic_1_1LinkEntryBroker.html',1,'ace_time::basic::LinkEntryBroker']]],
-  ['linkmanager_304',['LinkManager',['../classace__time_1_1LinkManager.html',1,'ace_time']]],
-  ['linkregistrar_305',['LinkRegistrar',['../classace__time_1_1basic_1_1LinkRegistrar.html',1,'ace_time::basic::LinkRegistrar'],['../classace__time_1_1extended_1_1LinkRegistrar.html',1,'ace_time::extended::LinkRegistrar']]],
-  ['linkregistrartemplate_306',['LinkRegistrarTemplate',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
-  ['linkregistrartemplate_3c_20basic_3a_3alinkentry_2c_20basic_3a_3alinkentrybroker_2c_20basic_3a_3alinkregistrybroker_20_3e_307',['LinkRegistrarTemplate&lt; basic::LinkEntry, basic::LinkEntryBroker, basic::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
-  ['linkregistrartemplate_3c_20extended_3a_3alinkentry_2c_20extended_3a_3alinkentrybroker_2c_20extended_3a_3alinkregistrybroker_20_3e_308',['LinkRegistrarTemplate&lt; extended::LinkEntry, extended::LinkEntryBroker, extended::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
-  ['linkregistrybroker_309',['LinkRegistryBroker',['../classace__time_1_1extended_1_1LinkRegistryBroker.html',1,'ace_time::extended::LinkRegistryBroker'],['../classace__time_1_1basic_1_1LinkRegistryBroker.html',1,'ace_time::basic::LinkRegistryBroker']]],
-  ['localdate_310',['LocalDate',['../classace__time_1_1LocalDate.html',1,'ace_time']]],
-  ['localdatetime_311',['LocalDateTime',['../classace__time_1_1LocalDateTime.html',1,'ace_time']]],
-  ['localtime_312',['LocalTime',['../classace__time_1_1LocalTime.html',1,'ace_time']]]
+  ['linkentrybroker_306',['LinkEntryBroker',['../classace__time_1_1extended_1_1LinkEntryBroker.html',1,'ace_time::extended::LinkEntryBroker'],['../classace__time_1_1basic_1_1LinkEntryBroker.html',1,'ace_time::basic::LinkEntryBroker']]],
+  ['linkmanager_307',['LinkManager',['../classace__time_1_1LinkManager.html',1,'ace_time']]],
+  ['linkregistrar_308',['LinkRegistrar',['../classace__time_1_1basic_1_1LinkRegistrar.html',1,'ace_time::basic::LinkRegistrar'],['../classace__time_1_1extended_1_1LinkRegistrar.html',1,'ace_time::extended::LinkRegistrar']]],
+  ['linkregistrartemplate_309',['LinkRegistrarTemplate',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
+  ['linkregistrartemplate_3c_20basic_3a_3alinkentry_2c_20basic_3a_3alinkentrybroker_2c_20basic_3a_3alinkregistrybroker_20_3e_310',['LinkRegistrarTemplate&lt; basic::LinkEntry, basic::LinkEntryBroker, basic::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
+  ['linkregistrartemplate_3c_20extended_3a_3alinkentry_2c_20extended_3a_3alinkentrybroker_2c_20extended_3a_3alinkregistrybroker_20_3e_311',['LinkRegistrarTemplate&lt; extended::LinkEntry, extended::LinkEntryBroker, extended::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
+  ['linkregistrybroker_312',['LinkRegistryBroker',['../classace__time_1_1extended_1_1LinkRegistryBroker.html',1,'ace_time::extended::LinkRegistryBroker'],['../classace__time_1_1basic_1_1LinkRegistryBroker.html',1,'ace_time::basic::LinkRegistryBroker']]],
+  ['localdate_313',['LocalDate',['../classace__time_1_1LocalDate.html',1,'ace_time']]],
+  ['localdatetime_314',['LocalDateTime',['../classace__time_1_1LocalDateTime.html',1,'ace_time']]],
+  ['localtime_315',['LocalTime',['../classace__time_1_1LocalTime.html',1,'ace_time']]]
 ];