+++ docs/html/search/classes_6.js
@@ -1,14 +1,14 @@
 var searchData=
 [
-  ['timeoffset_319',['TimeOffset',['../classace__time_1_1TimeOffset.html',1,'ace_time']]],
-  ['timeperiod_320',['TimePeriod',['../classace__time_1_1TimePeriod.html',1,'ace_time']]],
-  ['timezone_321',['TimeZone',['../classace__time_1_1TimeZone.html',1,'ace_time']]],
-  ['timezonedata_322',['TimeZoneData',['../structace__time_1_1TimeZoneData.html',1,'ace_time']]],
-  ['transitionresulttemplate_323',['TransitionResultTemplate',['../structace__time_1_1extended_1_1TransitionResultTemplate.html',1,'ace_time::extended']]],
-  ['transitionstoragetemplate_324',['TransitionStorageTemplate',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
-  ['transitionstoragetemplate_3c_20kmaxtransitions_2c_20extended_3a_3azoneerabroker_2c_20extended_3a_3azonepolicybroker_2c_20extended_3a_3azonerulebroker_20_3e_325',['TransitionStorageTemplate&lt; kMaxTransitions, extended::ZoneEraBroker, extended::ZonePolicyBroker, extended::ZoneRuleBroker &gt;',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
-  ['transitionstoragetemplate_3c_20kmaxtransitions_2c_20zeb_2c_20zpb_2c_20zrb_20_3e_326',['TransitionStorageTemplate&lt; kMaxTransitions, ZEB, ZPB, ZRB &gt;',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
-  ['transitiontemplate_327',['TransitionTemplate',['../structace__time_1_1basic_1_1TransitionTemplate.html',1,'ace_time::basic::TransitionTemplate&lt; ZIB, ZEB, ZPB, ZRB &gt;'],['../structace__time_1_1extended_1_1TransitionTemplate.html',1,'ace_time::extended::TransitionTemplate&lt; ZEB, ZPB, ZRB &gt;']]],
-  ['transitiontemplate_3c_20basic_3a_3azoneinfobroker_2c_20basic_3a_3azoneerabroker_2c_20basic_3a_3azonepolicybroker_2c_20basic_3a_3azonerulebroker_20_3e_328',['TransitionTemplate&lt; basic::ZoneInfoBroker, basic::ZoneEraBroker, basic::ZonePolicyBroker, basic::ZoneRuleBroker &gt;',['../structace__time_1_1basic_1_1TransitionTemplate.html',1,'ace_time::basic']]],
-  ['transitiontemplate_3c_20extended_3a_3azoneerabroker_2c_20extended_3a_3azonepolicybroker_2c_20extended_3a_3azonerulebroker_20_3e_329',['TransitionTemplate&lt; extended::ZoneEraBroker, extended::ZonePolicyBroker, extended::ZoneRuleBroker &gt;',['../structace__time_1_1extended_1_1TransitionTemplate.html',1,'ace_time::extended']]]
+  ['timeoffset_322',['TimeOffset',['../classace__time_1_1TimeOffset.html',1,'ace_time']]],
+  ['timeperiod_323',['TimePeriod',['../classace__time_1_1TimePeriod.html',1,'ace_time']]],
+  ['timezone_324',['TimeZone',['../classace__time_1_1TimeZone.html',1,'ace_time']]],
+  ['timezonedata_325',['TimeZoneData',['../structace__time_1_1TimeZoneData.html',1,'ace_time']]],
+  ['transitionresulttemplate_326',['TransitionResultTemplate',['../structace__time_1_1extended_1_1TransitionResultTemplate.html',1,'ace_time::extended']]],
+  ['transitionstoragetemplate_327',['TransitionStorageTemplate',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
+  ['transitionstoragetemplate_3c_20kmaxtransitions_2c_20extended_3a_3azoneerabroker_2c_20extended_3a_3azonepolicybroker_2c_20extended_3a_3azonerulebroker_20_3e_328',['TransitionStorageTemplate&lt; kMaxTransitions, extended::ZoneEraBroker, extended::ZonePolicyBroker, extended::ZoneRuleBroker &gt;',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
+  ['transitionstoragetemplate_3c_20kmaxtransitions_2c_20zeb_2c_20zpb_2c_20zrb_20_3e_329',['TransitionStorageTemplate&lt; kMaxTransitions, ZEB, ZPB, ZRB &gt;',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
+  ['transitiontemplate_330',['TransitionTemplate',['../structace__time_1_1basic_1_1TransitionTemplate.html',1,'ace_time::basic::TransitionTemplate&lt; ZIB, ZEB, ZPB, ZRB &gt;'],['../structace__time_1_1extended_1_1TransitionTemplate.html',1,'ace_time::extended::TransitionTemplate&lt; ZEB, ZPB, ZRB &gt;']]],
+  ['transitiontemplate_3c_20basic_3a_3azoneinfobroker_2c_20basic_3a_3azoneerabroker_2c_20basic_3a_3azonepolicybroker_2c_20basic_3a_3azonerulebroker_20_3e_331',['TransitionTemplate&lt; basic::ZoneInfoBroker, basic::ZoneEraBroker, basic::ZonePolicyBroker, basic::ZoneRuleBroker &gt;',['../structace__time_1_1basic_1_1TransitionTemplate.html',1,'ace_time::basic']]],
+  ['transitiontemplate_3c_20extended_3a_3azoneerabroker_2c_20extended_3a_3azonepolicybroker_2c_20extended_3a_3azonerulebroker_20_3e_332',['TransitionTemplate&lt; extended::ZoneEraBroker, extended::ZonePolicyBroker, extended::ZoneRuleBroker &gt;',['../structace__time_1_1extended_1_1TransitionTemplate.html',1,'ace_time::extended']]]
 ];