+++ docs/html/search/variables_c.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['startdatetime_577',['startDateTime',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#a9c77b33abfc02eb0b1450e7147ea692e',1,'ace_time::extended::MatchingEraTemplate::startDateTime()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#acf25fdb69df5d5f91c856f9e805c3792',1,'ace_time::extended::TransitionTemplate::startDateTime()']]],
-  ['startepochseconds_578',['startEpochSeconds',['../structace__time_1_1basic_1_1TransitionTemplate.html#a2a6fa21f53305bb789f0fa98c68db62f',1,'ace_time::basic::TransitionTemplate::startEpochSeconds()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#a490879be0a08acbf211c0a943d99c36e',1,'ace_time::extended::TransitionTemplate::startEpochSeconds()']]],
-  ['startyear_579',['startYear',['../structace__time_1_1internal_1_1ZoneContext.html#a7859ffdc83bb5269bd265079fcde20e8',1,'ace_time::internal::ZoneContext']]]
+  ['startdatetime_583',['startDateTime',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#a9c77b33abfc02eb0b1450e7147ea692e',1,'ace_time::extended::MatchingEraTemplate::startDateTime()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#acf25fdb69df5d5f91c856f9e805c3792',1,'ace_time::extended::TransitionTemplate::startDateTime()']]],
+  ['startepochseconds_584',['startEpochSeconds',['../structace__time_1_1basic_1_1TransitionTemplate.html#a2a6fa21f53305bb789f0fa98c68db62f',1,'ace_time::basic::TransitionTemplate::startEpochSeconds()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#a490879be0a08acbf211c0a943d99c36e',1,'ace_time::extended::TransitionTemplate::startEpochSeconds()']]],
+  ['startyear_585',['startYear',['../structace__time_1_1internal_1_1ZoneContext.html#a7859ffdc83bb5269bd265079fcde20e8',1,'ace_time::internal::ZoneContext']]]
 ];